third_party/blink/web_tests/external/wpt/html/webappapis/dynamic-markup-insertion/document-write/write-active-document-expected.txt

This is a testharness.js-based test.
FAIL document.write only writes to active documents assert_equals: expected (string) "IFRAME" but got (undefined) undefined
Harness: the test ran to completion.
